Improve waybar style and config
Added a top/bottom small border, changed the background color to be a
little blacker, and increased the refresh interval to 10 seconds minimum
to make waybar less CPU intensive.

Also improved the mpris module to display a specific icon when firefox
is playing sound, and the volume module to display the volume
percentage.