Update screenshot.sh and lockscreen.sh
Use new screen identifier (ignored in my setup by the lockscreen.sh
script) and simplify lockscreen.sh to rely on screenshot.sh.
screenshot.sh is also improved to add a mode to not copy the screenshot
to the clipboard (in this case, for use with the lockscreen) and to
allow specifying a custom screenshot path for the saved file. Other
small improvements also done (shellcheck, factorizations...)